Separated double and float.

+++ exposed-gradle/src/test/resources/databases/vartypes/FloatingPointTypes.kt
@@ -1,6 +1,6 @@
 object FloatingPointTypes : Table() {
     val f1: Column<Float> = float("f1")
-    val f2: Column<Float> = float("f2")
-    val f3: Column<Float> = float("f3")
-    val f4: Column<Float> = float("f4")
+    val f2: Column<Double> = double("f2")
+    val f3: Column<Double> = double("f3")
+    val f4: Column<Double> = double("f4")
 }
